[33기 이윤섭] MODIFIED: Main.js 내 피드 구조 리팩도링
