+++ src/pages/younseop/Main/Main.scss
@@ -47,6 +47,7 @@ $bold: 700;
         border: $border;
         background-color: #fff;
         border-radius: $radius-3;
+        margin-bottom: 24px;
 
         .postProfile {
           @include flexSpaceBetween;
